+++ src/DiagramElement/ArrowDependency.php
@@ -12,9 +12,19 @@ final class ArrowDependency extends Arrow
 
     public function toString(PhpClass $toClass): string
     {
-        if (strpos($this->getTo()->getName(), '[]') === false) {
-            return sprintf('  %s %s %s', $this->getFrom()->getClassNameAlias(), $this->figure, $toClass->getClassNameAlias());
+        if (strpos($this->getTo()->getName(), '[]') !== false) {
+            // ex. Product[]
+            return $this->getExpression($toClass);
         }
+        if (strpos($this->getTo()->getName(), 'array<') === 0) {
+            // ex. array<Product> or array<int, Product>
+            return $this->getExpression($toClass);
+        }
+        return sprintf('  %s %s %s', $this->getFrom()->getClassNameAlias(), $this->figure, $toClass->getClassNameAlias());
+    }
+
+    private function getExpression(PhpClass $toClass): string
+    {
         return sprintf('  %s "1" %s "*" %s', $this->getFrom()->getClassNameAlias(), $this->figure, str_replace('[]', '', $toClass->getClassNameAlias()));
     }
 }