+++ test/fixtures/array-expression-in-doc/product/Product.php
@@ -1,0 +1,15 @@
+<?php
+namespace hoge\fuga\product;
+
+class Product {
+    /** @var array<Tag> */
+    private array $tags = [];
+
+    /**
+     * @return array<Tag>
+     */
+    public function getTags(): array
+    {
+        return $this->tags;
+    }
+}